refactor(standalone): split tests into cross-platform and Unix-only files
Tests that use shell scripts (#!/bin/sh) as fake binaries, Unix PATH
separators, or process group management are moved to
standalone_unix_test.go with a //go:build !windows constraint.

The remaining tests in standalone_test.go run on all platforms.
Platform-specific test helpers (longSleepCommand, failCommand) are
in standalone_process_{unix,windows}_test.go.
